+++ src/constants/position.ts
@@ -1,0 +1,8 @@
+export type PositionType = {
+  [key: string]: string;
+};
+export const Position: PositionType = {
+  동아리원: 'MEMBER',
+  임원: 'EXECUTIVE',
+  회장: 'LEADER',
+};
