--- src/constants/qna.ts
@@ -1,49 +1,0 @@
-type QnAItem = {
-  Q: string;
-  A: string;
-};
-
-export const QnA: QnAItem[] = [
-  {
-    Q: '전동대회 참여가 어려울 경우 어떻게 해야 하나요?',
-    A: `동아리 대표자가 전동대회 참석이 어려울 경우, 위임장 제출 후 대리인이 전동대회에 참여할 수 있습니다.
-      ✔위임장은 공고된 개회 시간 전까지 서면 제출하셔야 합니다. 
-      ✔위임장은 공고된 개회 시간 전까지 서면 제출하셔야 합니다. 
-      ✔전동대회 3회 불참 시 주의 1회, 4회 불참 시 경고 1회이니 참고 부탁드립니다. \n
-      사전 위임장 제출은 총동아리연합회 이메일(mjuclubuth@gmail.com)으로 부탁드립니다.`,
-  },
-  {
-    Q: '활동 보고서 제출',
-    A: `활동보고서는 한 회차당 총 2개의 활동 보고서 제출이 가능합니다.활동보고서는 총동아리연합회 웹사이트 띵동 어드민 페이지에서 제출 가능합니다. 
-      *한 회차 = 총 2주를 의미 ex. 9회차 = 2024년 9월 2일 ~ 2024년 월 15일간 활동한 활동 2개에 관련한 보고서 제출 기한을 의미함. 
-      *동일한 날짜로 보고서 2개 제출은 불가합니다. (ex. 활동 1/ 2 모두 9월 2일에 이루어진 활동 => 1개만 점수 부여) 
-      *활동 사진에 날짜/시간 두가지 모두가 반드시 나와야 점수 인정이 가능합니다. 첨부드린 활동보고서 가이드라인을 준수하시어 제출 부탁드립니다. 
-      * 활동 사진 누락시, 해당 회차 활동보고서 인정되지 않습니다 ! 
-      * 활동보고서는 최소 인원은 '동아리원 회원 명단'에 등록되어 있는 동아리원 5인 이상 입니다 !!! 
-      * 웹사이트 띵동 공지사항에 올려드린 활동보고서 제출 달력 및 총동아리연합회 인스타그램을 통해 공지되는 총동아리연합회 달력에서 활동보고서 제출 기한 반드시 확인 부탁드립니다.`,
-  },
-  {
-    Q: 'PLU:th(페이백 제도)는 어떻게 신청하나요?',
-    A: `총동아리연합회 카카오톡 채널 / 인스타그램 링크트리를 통해페이백 신청 서류를 다운 받으신 후 작성하시어 총동아리연합회 이메일로 제출 부탁드립니다. 
-      5만원 한도 내에서는 횟수 제한이 없습니다. 한도 내에서 세부적으로 나누어 여러번 지원 받으셔도 무방합니다. 
-      ex. 3만원 /1만원/1만원 -> 총 세번 신청 = 가능`,
-  },
-  {
-    Q: '동아리 평가 기준이 있나요?',
-    A: `2024년 동아리 평가 기준은 다음 첨부 파일과 같습니다. \n 추가적으로 인스타그램 링크트리 / 띵동 웹사이트 공지사항/ 인스타그램 링크트리에서도 확인 가능합니다.`,
-  },
-  {
-    Q: '동아리 점수 확인은 어디서 가능한가요?',
-    A: `동아리 실시간 점수 확인은 띵동 어드민 웹사이트에서 확인 가능합니다. 등수의 경우, 매달 말 회장님께 개별적으로 전달드리고 있습니다.`,
-  },
-  {
-    Q: '준동아리 등록은 어떻게 하나요?',
-    A: `준동아리 가/신규등록 기한은 해당 년도에 배포되는 회칙에 의거하여 시행됩니다. \n 당해년도 총동아리연합회가 공지하는 사항을 따라주시면 되겠습니다. 
-      2024년 기준, 준동아리 가등록은 총동아리연합회 회칙에 의거하여 11월에 시행됩니다.`,
-  },
-  {
-    Q: '동아리방 수리 보수 신청',
-    A: `동아리방 수리 보수 신청은 총동아리연합회 웹사이트 띵동의'수리보수'탭에서 가능합니다. \n 수리보수가 필요한 시설물 사진을 찍어, 설명과 함께 제출해주시면 됩니다. 
-      - 교내 시설지원팀 관할에서 수리 보수가 이루어지기에, 실시간으로 수리보수가 이루어지지 않는 점 총동아리연합회 집행부가 수리 보수 신청 건은 상시 확인하고 있으며, 최대한 빠른 수리 보수가 이루어질 수 있도록 최선을 다하겠습니다.`,
-  },
-];
